web: move crm experience to dedicated route
